Add [UNFINISHED] basic support for png2latex.
This is used by creating a code block with language of !latex

This passes whatever is in the code block (beginning and ending whitespace stripped) to png2latex
and returning the output as an image. This allows you to do more advanced mathematics than mathjax
allows, as well as allowing you to use any LaTeX code.

This is probably NOT secure. It's definitely not secure against a DoS.